Moved to loader to component
Also changed the suggestion from skanehira
